Fix tests
- Move tests in chat_template.rs to tests/
